requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.1
Flask-Mail==0.9.1
python-dotenv==1.0.1
itsdangerous==2.2.0
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
flask-cors==4.0.0
gunicorn==22.0.0
psycopg[binary]==3.2.9
joblib==1.5.2
numpy==2.3.3
scikit-learn==1.7.2
